+++ src/common/data_structures.ts
@@ -6,11 +6,11 @@ import {
 } from '../../proto/world-engine.pb';
 import { HistoryItem } from '../components/history';
 import { AdditionalPhonemeInfo } from '../entities/inworld_packet.entity';
+import { SessionToken } from '../entities/session_token.entity';
 
 export interface Capabilities {
   audio?: boolean;
   emotions?: boolean;
-  continuation?: boolean;
   interruptions?: boolean;
   narratedActions?: boolean;
   phonemes?: boolean;
@@ -35,13 +35,6 @@ export interface User {
 
 export interface Client {
   id?: string;
-}
-
-export interface SessionToken {
-  token: string;
-  type: string;
-  expirationTime: string;
-  sessionId: string;
 }
 
 export interface Gateway {