+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [Unreleased]
+
+### Added
+
+- Scene name pattern validation
+- Allow to get the conversation state
+- Allow to propagate previous conversation state to new connection
+
 ## [1.8.0] - 2023-09-29
 
 ### Added